requirements.txt
streamlit 
langchain 
langchain-community
openai 
tiktoken
reportlab 
python-dotenv
gTTS
matplotlib